MOD: #8.15 Edit Profile, Change Password, Upload Video

+++ src/controller/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 export const trending = async (req, res) => {
     try {
@@ -11,11 +12,11 @@ export const trending = async (req, res) => {
 
 export const watch = async (req, res) => {
     const { id } = req.params;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate('owner');
     if (!video) {
         return res.status(404).render('404', { pageTitle: 'Video Not Found' });
     }
-    return res.render('watch', { pageTitle: video.title, video });
+    return res.render('video/watch', { pageTitle: video.title, video });
 }
 
 export const getEdit = async (req, res) => {
@@ -23,6 +24,9 @@ export const getEdit = async (req, res) => {
     const video = await Video.findById(id);
     if (!video) {
         return res.status(404).render('404', { pageTitle: 'Video Not Found' });
+    }
+    if (String(video.owner) !== String(req.session.user._id)) {
+        return res.status(403).redirect('/');
     }
     return res.render('Edit', {
         pageTitle: `Editing ${video.title}`,
@@ -32,10 +36,13 @@ export const getEdit = async (req, res) => {
 
 export const postEdit = async (req, res) => {
     const { id } = req.params;
-    const {title, description, hashtags} = req.body;
+    const { title, description, hashtags } = req.body;
     const video = await Video.exists({ _id: id });
     if (!video) {
         return res.status(404).render('404', { pageTitle: 'Video Not Found' });
+    }
+    if (String(video.owner) !== String(req.session.user._id)) {
+        return res.status(403).redirect('/');
     }
     await Video.findByIdAndUpdate(id, {
         title,
@@ -46,22 +53,28 @@ export const postEdit = async (req, res) => {
 }
 
 export const getUpload = (req, res) => {
-    return res.render('upload', { pageTitle: `Upload Video` });
+    return res.render('video/upload', { pageTitle: `Upload Video` });
 }
 
 export const postUpload = async (req, res) => {
     try {
+        const { user: { _id: owner } } = req.session;
+        const { path: fileUrl } = req.file;
         const { title, description, hashtags } = req.body;
-        const saveRes = await Video.create({
+        const newVideo = await Video.create({
+            fileUrl,
             title,
             description,
+            owner,
             hashtags: Video.formatHashtags(hashtags),
         });
-        console.log(saveRes);
+        const user = await User.findById(owner);
+        user.videos.push(newVideo._id);
+        await user.save();
         return res.redirect('/');
     } catch (error) {
         console.error('error => ', error);
-        return res.render('upload', {
+        return res.render('video/upload', {
             pageTitle: `Upload Video`,
             errorMessage: error?._message
         });
@@ -71,6 +84,13 @@ export const postUpload = async (req, res) => {
 export const remove = async (req, res) => {
     try {
         const { id } = req.params;
+        const video = await Video.exists({ _id: id });
+        if (!video) {
+            return res.status(404).render('404', { pageTitle: 'Video Not Found' });
+        }
+        if (String(video.owner) !== String(req.session.user._id)) {
+            return res.status(403).redirect('/');
+        }
         await Video.findByIdAndDelete(id);
         return res.redirect('/');
     } catch (error) {
